Add docstings and type annotations
Add docstings and type annotations to all Matrix class methods
